add scroll directive
