+++ src/app/features-modules/public/landing-page/components/banner/banner.component.html
@@ -4,14 +4,13 @@
     <img class="absolute -right-20 opacity-40 top-0" src="assets/banner/right-shape.png" alt="">
     <div class="container">
         <div class="w-full flex flex-col h-screen justify-center">
-            <div class="w-full  flex justify-between items-center gap-8">
+            <div class="w-full  flex justify-between items-center gap-8 mt-20">
                 <div class="w-[65%] flex flex-col">
                     <div class="w-full sm:w-full space-y-12 mt-8 md:mt-0">
                         <div class="banner-left-top relative">
                             <h1 class="leading-[1.2] md:text-[44px] drop-shadow-md 2xl:text-[58px] font-Roboto font-extrabold text-white tracking-[1.2px]">
-                                Consult <span class="text-primary ">Best Doctors</span> from  Your Home</h1>
-                            <p class="mt-4 md:text-[14px] 2xl:text-[16px] font-poppins font-medium text-gray-200">Lorem ipsum dolor sit amet, consectetur adipiscing elit, Lorem ipsum dolor sit amet.
-                                lorm10 Lorem ipsum.
+                                Your All-In-One <span class="text-primary ">Health</span> Platform</h1>
+                            <p class="mt-4 md:text-[15px] 2xl:text-[17px] leading-[30px] font-medium text-secondary-dark">Discover a comprehensive healthcare solution right at your fingertips. Consult with a qualified medical professional from anywhere, at any time, and receive instant prescriptions. Soowgood is dedicated to ensuring your health and wellness.
                             </p>
                         </div>
                         <!-- <div class="flex gap-4 flex-wrap max-w-[580px] 2xl:mt-12">
