+++ src/app/features-modules/public/landing-page/components/client-feedback/client-feedback.component.html
@@ -8,34 +8,67 @@
             </p>
         </div>
         <div class="mt-16 grid grid-cols-3 gap-5">
-            <div class="p-6 bg-primary rounded-lg">
-                <p class="font-Roboto text-[16px] text-gray-400">Lorem ipsum dolor sit, amet consectetur adipisicing
+            <div class="p-6 bg-primary/5 border-secondary-dark/20 border-[1px] rounded-lg">
+                <p class="font-Roboto text-[16px] text-primary">Lorem ipsum dolor sit, amet consectetur adipisicing
                     elit. Laboriosam,
                     voluptas. Minus soluta expedita excepturi, libero quas rerum labore ad molestiae?
                 </p>
-                <div class="mt-4 flex gap-5">
-                    <img class="w-[60px] h-[60px] rounded-full bg-white" src="" alt="">
-                    <h2 class="font-Roboto text-[20px] font-semibold text-secondary">Patient name</h2>
+                <div class="mt-4 flex gap-5 ">
+                    <img class="w-[60px] border-secondary/50 border-4 h-[60px] rounded-full bg-white" src="assets/patients/p1.webp" alt="patient">
+                 <div>
+
+                     <h2 class="font-Roboto text-[20px] font-semibold text-primary">Patient name</h2>
+                     <div class="mt-1 flex gap-1">
+                         <i class="fa-solid text-sm fa-star text-secondary"></i>
+                         <i class="fa-solid text-sm fa-star text-secondary"></i>
+                         <i class="fa-solid text-sm fa-star text-secondary"></i>
+                         <i class="fa-solid text-sm fa-star text-secondary"></i>
+                         <i class="fa-solid text-sm  fa-star text-secondary"></i>
+                         <p class="text-sm">(5 start)</p>
+                     </div>
+                 </div>
                 </div>
             </div>
-            <div class="p-6 bg-primary rounded-lg">
-                <p class="font-Roboto text-[16px] text-gray-400">Lorem ipsum dolor sit, amet consectetur adipisicing
+            <div class="p-6 bg-primary/5 border-secondary-dark/20 border-[1px] rounded-lg">
+                <p class="font-Roboto text-[16px] text-primary">Lorem ipsum dolor sit, amet consectetur adipisicing
                     elit. Laboriosam,
                     voluptas. Minus soluta expedita excepturi, libero quas rerum labore ad molestiae?
                 </p>
-                <div class="mt-4 flex gap-5">
-                    <img class="w-[60px] h-[60px] rounded-full bg-white" src="" alt="">
-                    <h2 class="font-Roboto text-[20px] font-semibold text-secondary">Patient name</h2>
+                <div class="mt-4 flex gap-5 ">
+                    <img class="w-[60px] border-secondary/50 border-4 h-[60px] rounded-full bg-white" src="assets/patients/p1.webp" alt="patient">
+                 <div>
+
+                     <h2 class="font-Roboto text-[20px] font-semibold text-primary">Patient name</h2>
+                     <div class="mt-1 flex gap-1">
+                         <i class="fa-solid text-sm fa-star text-secondary"></i>
+                         <i class="fa-solid text-sm fa-star text-secondary"></i>
+                         <i class="fa-solid text-sm fa-star text-secondary"></i>
+                         <i class="fa-solid text-sm fa-star text-secondary"></i>
+                         <i class="fa-solid text-sm  fa-star text-secondary"></i>
+                         <p class="text-sm">(5 start)</p>
+                     </div>
+                 </div>
                 </div>
             </div>
-            <div class="p-6 bg-primary rounded-lg">
-                <p class="font-Roboto text-[16px] text-gray-400">Lorem ipsum dolor sit, amet consectetur adipisicing
+            <div class="p-6 bg-primary/5 border-secondary-dark/20 border-[1px] rounded-lg">
+                <p class="font-Roboto text-[16px] text-primary">Lorem ipsum dolor sit, amet consectetur adipisicing
                     elit. Laboriosam,
                     voluptas. Minus soluta expedita excepturi, libero quas rerum labore ad molestiae?
                 </p>
-                <div class="mt-4 flex gap-5">
-                    <img class="w-[60px] h-[60px] rounded-full bg-white" src="" alt="">
-                    <h2 class="font-Roboto text-[20px] font-semibold text-secondary">Patient name</h2>
+                <div class="mt-4 flex gap-5 ">
+                    <img class="w-[60px] border-secondary/50 border-4 h-[60px] rounded-full bg-white" src="assets/patients/p1.webp" alt="patient">
+                 <div>
+
+                     <h2 class="font-Roboto text-[20px] font-semibold text-primary">Patient name</h2>
+                     <div class="mt-1 flex gap-1">
+                         <i class="fa-solid text-sm fa-star text-secondary"></i>
+                         <i class="fa-solid text-sm fa-star text-secondary"></i>
+                         <i class="fa-solid text-sm fa-star text-secondary"></i>
+                         <i class="fa-solid text-sm fa-star text-secondary"></i>
+                         <i class="fa-solid text-sm  fa-star text-secondary"></i>
+                         <p class="text-sm">(5 start)</p>
+                     </div>
+                 </div>
                 </div>
             </div>
         </div>
