+++ src/app/features-modules/public/landing-page/components/live-doctor-card/live-doctor-card.component.html
@@ -2,7 +2,7 @@
     <img class="absolute  card-img" src="assets/banner/shape-24.jpg" alt="">
     <div class="absolute top-0 left-0 w-full h-full z-10">
         <div class="w-[90px] relative after:w-[15px] after:h-[15px] after:border-[2px]  after:bg-green-400 after:rounded-full after:absolute after:bottom-0 after:right-0 ml-6 mt-6 overflow-hidden p-1 h-[90px] rounded-[10px] border-[1px] bg-secondary/10">
-          <img class="object-cover rounded-lg w-full h-full" src="assets/doctor/doctor.jpg" alt="">
+          <img class="object-cover rounded-lg w-full h-full" src="assets/doctor/dr.jpeg" alt="">
         
         </div>
         <div>
