+++ src/app/features-modules/public/landing-page/components/live-doctors/live-doctors.component.html
@@ -13,7 +13,7 @@
 
                 </div>
                 <div class="flex gap-5">
-                    <button (click)="onClickSeeMore('livedoctors')">see more</button>
+                   
                     <button class="btn-secondary"><i class="fa-solid fa-arrow-left"></i></button>
                     <button class="btn-secondary"><i class="fa-solid fa-arrow-right"></i></button>
                 </div>
@@ -30,6 +30,7 @@
                 <app-live-doctor-card *ngFor="let details of liveDoctorList" [doctorDetails]="details"></app-live-doctor-card>
 
             </div>
+            <button class="rounded-3xl mt-10 px-6 py-2 bg-secondary text-white mx-auto" (click)="onClickSeeMore('livedoctors')">see more</button>
         </div>
     </div>
 </div>
