+++ src/app/features-modules/public/landing-page/components/signup-marketing-banner/signup-marketing-banner.component.html
@@ -1,8 +1,8 @@
-<section class="bg-primary">
+<section class="bg-primary rounded-t-[70px]">
     <div class="grid max-w-screen-xl px-4 py-8 mx-auto lg:gap-8 xl:gap-0 lg:py-16 lg:grid-cols-12">
         <div class="mr-auto place-self-center lg:col-span-7">
-            <h1 class="max-w-xl mb-4 text-[44px] font-bold font-Roboto text-secondary tracking-tigh">Become a Doctor</h1>
-            <p class="max-w-2xl mb-6 font-light text-gray-300 lg:mb-8 md:text-lg lg:text-xl dark:text-gray-400">From checkout to global sales tax compliance, companies around the world use Flowbite to simplify their payment stack.</p>
+            <h1 class="max-w-xl mb-4 text-[44px] font-bold font-Roboto text-secondary tracking-tigh">Consult your patients anywhere anytime.</h1>
+            <p class="max-w-2xl mb-6 font-light text-gray-300 lg:mb-8 md:text-lg lg:text-xl dark:text-gray-400">Unable to reach your clinic/chamber on time? With Soowgood, you can now provide consultations to your patients from anywhere, at any time. Join as a registered doctor and begin consultations instantly!"</p>
             <a href="#" class="inline-flex items-center justify-center px-5 py-3 mr-3 text-base font-medium text-center text-white rounded-lg bg-primary-700 hover:bg-primary-800 focus:ring-4 focus:ring-primary-300 dark:focus:ring-primary-900">
                 Get started
                 <svg class="w-5 h-5 ml-2 -mr-1" fill="currentColor" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" d="M10.293 3.293a1 1 0 011.414 0l6 6a1 1 0 010 1.414l-6 6a1 1 0 01-1.414-1.414L14.586 11H3a1 1 0 110-2h11.586l-4.293-4.293a1 1 0 010-1.414z" clip-rule="evenodd"></path></svg>
@@ -12,7 +12,7 @@
             </a> 
         </div>
         <div class="hidden lg:mt-0 lg:col-span-5 lg:flex">
-            <img src="assets/banner/bcome.png" alt="mockup">
+            <img src="assets/doctor/online-dr.png" alt="mockup">
         </div>                
     </div>
 </section>
