+++ src/app/features-modules/public/landing-page/landing-page.component.html
@@ -1,9 +1,14 @@
+<div #scrolltop></div>
 <div class="relative">
     <app-banner></app-banner>
     <app-live-doctors></app-live-doctors>
-    <app-signup-marketing-banner></app-signup-marketing-banner>
+    <app-app-marketing></app-app-marketing>
     <app-client-feedback></app-client-feedback>
     <app-pricing></app-pricing>
-    <app-app-marketing></app-app-marketing>
+    <app-signup-marketing-banner></app-signup-marketing-banner>
     <app-newsletter></app-newsletter>
 </div>  
+
+<div  appScroll class="absolute bottom-10 right-10 cursor-pointer bg-secondary p-4 text-white rounded-md">
+    <i class="fa-solid fa-arrow-up"></i>
+</div>
