+++ src/app/features-modules/public/landing-page/landing-page.module.ts
@@ -18,7 +18,8 @@ import { AppMarketingComponent } from './components/app-marketing/app-marketing.
 import { ClientFeedbackComponent } from './components/client-feedback/client-feedback.component';
 import { NewsletterComponent } from './components/newsletter/newsletter.component';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
-import { CacheInterceptor } from 'src/app/shared/utils/CacheInterceptor';
+import { CacheInterceptor } from 'src/app/shared/utils/interceptors/CacheInterceptor';
+import { ScrollDirective } from 'src/app/shared/directive/scroll/scroll.directive';
 
 const routes: Route[] = [
   {
@@ -41,6 +42,7 @@ const routes: Route[] = [
     AppMarketingComponent,
     ClientFeedbackComponent,
     NewsletterComponent,
+    ScrollDirective
   ],
   imports: [CommonModule,
     RouterModule.forChild(routes),
