+++ src/app/layout-module/header-footer/footer/footer.component.html
@@ -1,48 +1,49 @@
 <footer class="bg-primary">
     <div class="container footer py-16 text-white">
-            <div class="w-[320px]">
-                <img src="../../../../assets/SoowGood-Logo.png" alt="">
-                <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Aliquam, ratione. Beatae magni ipsa dicta officiis!</p>
-            </div> 
-            <div>
-              <span class="footer-title">Services</span> 
-              <a class="link link-hover">Branding</a> 
-              <a class="link link-hover">Design</a> 
-              <a class="link link-hover">Marketing</a> 
-              <a class="link link-hover">Advertisement</a>
-            </div> 
-            <div>
-              <span class="footer-title">Company</span> 
-              <a class="link link-hover">About us</a> 
-              <a class="link link-hover">Contact</a> 
-              <a class="link link-hover">Jobs</a> 
-              <a class="link link-hover">Press kit</a>
-            </div> 
-            <div>
-              <span class="footer-title">Legal</span> 
-              <a class="link link-hover">Terms of use</a> 
-              <a class="link link-hover">Privacy policy</a> 
-              <a class="link link-hover">Cookie policy</a>
-            </div>
-            <div>
-                <span class="footer-title">Find Us</span> 
-                <a class="link link-hover">Terms of use</a> 
-                <a class="link link-hover">Privacy policy</a> 
-                <a class="link link-hover">Cookie policy</a>
-              </div>
+        <div class="max-w-[220px]">
+            <img src="../../../../assets/SoowGood-Logo.png" alt="">
+            <p>Consult your patients anywhere anytime.</p>
+
+            <p class="mt-5"> <i class="fa-solid fa-phone mr-2 text-secondary "></i>+8801605-144633</p>
+            <p> <i class="fa-solid fa-paper-plane  mr-2 text-secondary"></i>info@soowgood.com</p>
+            <p class=" flex items-center"> <i class="fa-solid fa-location-dot mr-2 text-secondary"></i> <span>Mahtab
+                    Center (12th Floor), Room #10, Bijoynagar, Dhaka 1000</span></p>
+        </div>
+        <div>
+            <span class="font-Roboto font-semibold mb-4 text-[20px] text-white">Company</span>
+            <a class="link link-hover opacity-80">About us</a>
+            <a class="link link-hover opacity-80">Support</a>
+            <a class="link link-hover opacity-80">Soowgood points</a>
+            <a class="link link-hover opacity-80">Village booths</a>
+            <a class="link link-hover opacity-80">Press kit</a>
+        </div>
+        <div>
+            <span class="font-Roboto font-semibold mb-4 text-[20px] text-white">Platform</span>
+            <a class="link link-hover opacity-80">Branding</a>
+            <a class="link link-hover opacity-80">Design</a>
+            <a class="link link-hover opacity-80">Marketing</a>
+            <a class="link link-hover opacity-80">Advertisement</a>
+        </div>
+        <div>
+            <span class="font-Roboto font-semibold mb-4 text-[20px] text-white">Licence</span>
+            <a class="link link-hover opacity-80">TL : 039455/2022</a>
+            <a class="link link-hover opacity-80">BIN : 005606801-0208</a>
+            <a class="link link-hover opacity-80">Privacy policy</a>
+            <a class="link link-hover opacity-80">Terms & Condition</a>
+        </div>
+
 
     </div>
 </footer>
 <footer class="bg-primary">
-    <div class="container footer py-4 border-t text-white border-base-300">
+    <div class="container">
+        <img src="assets/payment/sslcommerz-banner.webp" alt="ssl-banner">
+    </div>
+    <div class="container mt-5 footer py-4 border-t text-white border-base-300">
+
         <div class="items-center grid-flow-col">
-            <svg width="24" height="24" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg" fill-rule="evenodd"
-                clip-rule="evenodd" class="fill-current">
-                <path
-                    d="M22.672 15.226l-2.432.811.841 2.515c.33 1.019-.209 2.127-1.23 2.456-1.15.325-2.148-.321-2.463-1.226l-.84-2.518-5.013 1.677.84 2.517c.391 1.203-.434 2.542-1.831 2.542-.88 0-1.601-.564-1.86-1.314l-.842-2.516-2.431.809c-1.135.328-2.145-.317-2.463-1.229-.329-1.018.211-2.127 1.231-2.456l2.432-.809-1.621-4.823-2.432.808c-1.355.384-2.558-.59-2.558-1.839 0-.817.509-1.582 1.327-1.846l2.433-.809-.842-2.515c-.33-1.02.211-2.129 1.232-2.458 1.02-.329 2.13.209 2.461 1.229l.842 2.515 5.011-1.677-.839-2.517c-.403-1.238.484-2.553 1.843-2.553.819 0 1.585.509 1.85 1.326l.841 2.517 2.431-.81c1.02-.33 2.131.211 2.461 1.229.332 1.018-.21 2.126-1.23 2.456l-2.433.809 1.622 4.823 2.433-.809c1.242-.401 2.557.484 2.557 1.838 0 .819-.51 1.583-1.328 1.847m-8.992-6.428l-5.01 1.675 1.619 4.828 5.011-1.674-1.62-4.829z">
-                </path>
-            </svg>
-            <p>SOOWGOOD Ltd. <br />Providing reliable tech since 1992</p>
+            <img src="assets/SoowGood-Icon.png" width="30px" alt="">
+            <p>SOOWGOOD Ltd. <br />Providing reliable helth solution since 2021</p>
         </div>
         <div class="md:place-self-center md:justify-self-end">
             <div class="grid grid-flow-col gap-4">
@@ -67,4 +68,5 @@
             </div>
         </div>
     </div>
+    \
 </footer>
