+++ src/app/layout-module/layout.module.ts
@@ -13,6 +13,7 @@ import { PublicLayoutTwoComponent } from './layouts/public-layout-two/public-lay
 
 import { PaymentSuccessComponent } from '../shared/components/payment-success/payment-success.component';
 import { isAuth } from '../auth-gurd/auth.service';
+import { ScrollDirective } from '../shared/directive/scroll/scroll.directive';
 
 const routes: Route[] = [
   {
@@ -126,6 +127,7 @@ const routes: Route[] = [
     AgentLayoutComponent,
     HeaderComponent,
     FooterComponent,
+    ScrollDirective,
     // DashboardHeaderComponent,
     PublicLayoutTwoComponent
   ],
