--- src/app/shared/directive/loading.directive.spec.ts
@@ -1,8 +1,0 @@
-import { LoadingDirective } from './loading.directive';
-
-describe('LoadingDirective', () => {
-  it('should create an instance', () => {
-    const directive = new LoadingDirective();
-    expect(directive).toBeTruthy();
-  });
-});
