--- src/app/shared/directive/loading.directive.ts
@@ -1,21 +1,0 @@
-import { Directive, Input, TemplateRef, ViewContainerRef } from '@angular/core';
-
-@Directive({
-  selector: '[appLoading]'
-})
-export class LoadingDirective {
-
-  constructor(
-    private templateRef: TemplateRef<any>,
-    private viewContainer: ViewContainerRef
-  ) {}
-
-  @Input() set appLoading(condition: boolean) {
-    if (condition) {
-      this.viewContainer.createEmbeddedView(this.templateRef);
-    } else {
-      this.viewContainer.clear();
-    }
-  }
-
-}
