+++ src/app/shared/directive/scroll/scroll.directive.spec.ts
@@ -1,0 +1,8 @@
+import { ScrollDirective } from './scroll.directive';
+
+describe('ScrollDirective', () => {
+  it('should create an instance', () => {
+    const directive = new ScrollDirective();
+    expect(directive).toBeTruthy();
+  });
+});
