+++ src/app/shared/directive/scroll/scroll.directive.ts
@@ -1,0 +1,10 @@
+import { Directive,  HostListener } from '@angular/core';
+
+@Directive({
+  selector: '[appScroll]'
+})
+export class ScrollDirective {
+  @HostListener('click') onClick() {
+    window.scrollTo({ top: 0, behavior: 'smooth' });  }
+
+}
