+++ src/app/shared/utils/interceptors/SplashInterceptor.ts
@@ -1,0 +1,29 @@
+
+import { Injectable } from '@angular/core';
+import { HttpInterceptor, HttpRequest, HttpHandler, HttpResponse } from '@angular/common/http';
+import { of } from 'rxjs';
+import { tap } from 'rxjs/operators';
+
+@Injectable()
+export class SplashInterceptor implements HttpInterceptor {
+  private cache = new Map<string, any>();
+
+  intercept(req: HttpRequest<any>, next: HttpHandler) {
+    if (req.method !== 'GET') {
+      return next.handle(req);
+    }
+
+    const cachedResponse = this.cache.get(req.url);
+    if (cachedResponse) {
+      return of(new HttpResponse({ body: cachedResponse }));
+    }
+
+    return next.handle(req).pipe(
+      tap(event => {
+        if (event instanceof HttpResponse) {
+          this.cache.set(req.url, event.body);
+        }
+      })
+    );
+  }
+}
