+++ src/app/shared/utils/interceptors/splash-screen.service.ts
@@ -1,0 +1,18 @@
+import { Injectable } from '@angular/core';
+import { BehaviorSubject } from 'rxjs';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class SplashScreenService {
+
+  public isLoading = new BehaviorSubject<boolean>(false);
+
+  showLoader(): void {
+    this.isLoading.next(true);
+  }
+
+  hideLoader(): void {
+    this.isLoading.next(false);
+  }
+}
