+++ src/index.html
@@ -23,7 +23,7 @@
     }
   </style>
 </head>
-<body >
+<body class="relative">
   
   <app-root></app-root>
