+++ src/styles.scss
@@ -32,7 +32,7 @@
     @apply flex items-center justify-between w-full ;
   }
   .header-logo {
-    @apply w-[220px];
+    @apply w-[240px];
   }
   .btn-primary {
     @apply active:scale-90 transition-all duration-75 h-fit text-white bg-gradient-to-r from-cyan-500 to-blue-500 hover:bg-gradient-to-bl focus:ring-4 focus:outline-none focus:ring-cyan-300  font-medium rounded-sm text-sm px-5 py-2 text-center mr-2;
